2021 day 14; part 2 ; ruby
YIKES. So much yikes

Had a hell of a hard time getting the mutable state correct
for managing the pairs of nucleotides. And then there
was the surprise fact that my count is off by 1 for some
reason that I still do not understand. There must be something
funky with how I am counting the values but I could not say how.