2022; ruby ; day 11 ; part 2
Holy crap.

Ran into a few problems here. Performance for one thing,
the loop turned into crawl around 700 iterations. Looking
at the performance of the code, it appears that the problem was
that we had very large numbers (BigNum) and that calling #to_s
on them would take increasingly long. This required parsing the
expressions properly, not using eval, and making the memory usage
and runtime complexity for this much simpler.

After that, we had warnings with some multiplication operations being
too large for ruby to handle natively. The problem refers to the need to
'find a creative way to keep your worry levels managable'. The logical way
to do this is to take the worry levels modulo the least common multiple
of all the divisors, since that does not change the modulo operation.

Phew. Good stuff.