2022; elixir; day 01; part 2
Ah yes, that concludes a full first day of AOC in Elixir.
Quite happy with this and glad that I managed to persevere. Required
a decent amount of research, trial and error, documentation reading,
and going to consult ChatGPT a decent amount.

Shocker, ChatGPT is NOT that useful for elixir. They hallucinate answers
frequently enough and give incorrect advice for certain methods.