2023; elixir; day 08; part 2
Added a basic maths module to handle GCD and LCM because
those come up all the dang time in AoC. Quite happy with the
refactoring that came about to re-use the walk function here
even though when I originally wrote this I needed to duplicate it
just to make sense of the logic.

Not visible : the time I spent trying to write a #walk_simultaneously
function that would take a list of locations and try to walk all of them
simultaneously, that was some real nonsense. Should have trusted the hint
when it said that it's going to take <b>significantly more steps<b> to escape.