2023; elixir; day 10; part 1
A bit hairy, and a lot of cases to consider.
There's got to be a better way to handle the connections from
one pipe to another. Perhaps specifying for each where they can
connect, and then checking to see if the connections do line up
from neighbor to neighbor are possible ?