2023; elixir; day 10; part 2
A real doozy of a doozy.

Did not expect ray tracing for the part two here. Luckily I recall seeing some
similar problems in the past, and was able to determine with some light googling
that raytracing is a valid solution here.

Sadly I had to hack this up a bit. The real tricky part is determining when you've crossed
a line and when you haven't.

For example, below you clearly have not crossed a line

.F-----7.

because you only really skirt the edge of the polygon. However, in this example

.F-----J

you clearly do cross a line because the F is going up-right and the J is going left-up

Now the tricky part is that S can play either role. It can be the top left corner

.S-----7.

and it can also be part of a vertical line with a horizontal jag across it

.S.....J

So ... without a sort of code to peek around and determine for sure
which direction the S is going, I can't generalize the current solution.