2023; elixir; day 16; part 1
Got stuck for a long while on a solution that still seems
like it should work, but gets an answer that is half as large as expected.

My first approach was to take each light beam and

* move it one square according to its direction
* decide which direction it should change (if at all)

Clearly that doesn't work, probably because of some edge case where
there are mirrors next to mirrors (or splitters).