2023; elixir; day 17; part 1
A painful one. Took a good amount of time to see
that there were several bugs in this. Notably in how I was calculating
the heat lost between two points, and also in remembering that we can arrive
in a square from several different directions, and we need to be able to remember that
so that we can re-visit a square a second time, if it means we have a lower
amount of heat lost the second time