2023; elixir; day 17; part 2
Still could be optimized better. Probably wasting a lot of
time deciding which of the options to pursue at the top of
the a_star function. Using a list there is probably far from
the most optimal choice, since we need to traverse this entire
list the entire time, even though we know these scores and
there were only a few insert / updates of the fScore map
in the previous iteration