2023; elixir; day 18; part 1
Got stuck for a long while trying to implement a ray-casting algorithm
again, just like day 10. Unlike day 10 I couldn't work around the fact that the
dang corners and bottom edges of the polygon produces areas where it is functionally
impossible to determine whether you are about to cross into the polygon or just touch
the tip of the outside of the polygon without entering at all.

Turns out that the shoelace formula for determining the area of a polygon, and that is
significantly faster and more reliable.

https://en.wikipedia.org/wiki/Shoelace_formula