2024; elixir; day 8 ; part 1
Lost some time because of some bad parsing, I had forgotten to trim
trailing whitespace when parsing vertical lines, doing this resulted in an
incorrect maximum y_index for the grid, which resulted in not enough antennae
being filtered out for being outside of the grid. Welp. Live and learn.