2024; elixir; day 9; part 1
Tremendously slow, this takes approximately 55 seconds to run on my m2 pro.
I suspect what is slow is that it's trying to run a bunch of slow list operations
and always re-iterating the list from scratch is causing us to pay a tremendous price
in a language that punishes you for operating on something that is not the head of a list.