2024; elixir; day 10; part 2
I'm somehow surprised this worked. I ended up writing the code during part 1 to find all of the unique paths
and ended up throwing it away because I did not know if I would need it. It turned out to be
hard to write it again and while doing so I discovered that in fact I did not need to keep track
of the trails as we walked them, just increment a sum as we walk the topographical map.

This was a nice little bit of code today. Feeling good.