2024; elixir; day 11; part 2
It took A LOT OF WORK to identify how to make this fast enough. This was not a good day for functional programming, as keeping the mutable
state around as a cache was a lot of work. Luckily throwing it in a separate process seemed to make this fast enoguh