2024; elixir; day 15; part 1
A bit of work to get all of the fiddly bits right, with respect to finding which boxes needed to be moved, etc.
Definitely worth extracting some common helpers for this kind of work, because
I feel like this sort of problem does appear at least once per year, sometimes
more than once.