2024; elixir; day 16; part 2
Ooof. A tough day. Lots of brain juice spilled over
debugging an A* algorithm. Turns out the hardest part
of this for me was to ensure that I wasn't making it harder
for myself than I needed to -- keeping a MapSet for the open set
would have been a good idea.